requirements.txt
aws-cdk-lib>=2.96.0
constructs>=10.3.0
aws-cdk.aws-lambda-python-alpha>=2.96.0a0
boto3>=1.28.62
pytest>=7.4.0
